+++ testsuite/tests/singlecluster/extensions/pipeline_policy/conftest.py
@@ -4,9 +4,6 @@ import pytest
 
 from openshift_client import OpenShiftPythonException
 
-from testsuite.kubernetes import Selector
-from testsuite.kubernetes.deployment import Deployment
-from testsuite.kubernetes.service import Service, ServicePort
 from testsuite.kuadrant.extensions.pipeline_policy import PipelinePolicy
 
 
@@ -23,37 +20,6 @@ def check_pipeline_policy_crd(cluster, skip_or_fail):
 def pipeline_policy(cluster, blame, route):
     """PipelinePolicy targeting the test HTTPRoute"""
     return PipelinePolicy.create_instance(cluster, blame("pipeline"), route)
-
-
-@pytest.fixture(scope="module")
-def threat_assessment_service(request, cluster, blame, module_label, testconfig):
-    """Deploys the ThreatAssessmentService gRPC backend"""
-    name = blame("threat")
-    match_labels = {"app": module_label, "deployment": name}
-
-    deployment = Deployment.create_instance(
-        cluster,
-        name,
-        container_name="threat-assessment",
-        image=testconfig["pipeline_policy_extension_service"]["image"],
-        ports={"grpc": 8080},
-        selector=Selector(matchLabels=match_labels),
-        labels={"app": module_label},
-    )
-    request.addfinalizer(deployment.delete)
-    deployment.commit()
-    deployment.wait_for_ready()
-
-    service = Service.create_instance(
-        cluster,
-        name,
-        selector=match_labels,
-        ports=[ServicePort(name="grpc", port=8080, targetPort="grpc")],
-        labels={"app": module_label},
-    )
-    request.addfinalizer(service.delete)
-    service.commit()
-    return service
 
 
 @pytest.fixture(scope="module", autouse=True)
